+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Do not force to return focus to the calendar button after the picker closes.
+
 
 4.2.0 - (May 21, 2019)
 ------------------